pyproject.toml
[project]
name = "python-coding-challenges"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.0",
    "ptw>=1.0.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-watch>=4.2.0",
    "pyyaml>=6.0.3",
    "uvicorn>=0.38.0",
]
[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-watch"
]
